Work on cat tile updating system (still broken)

+++ include/datatypes/cat.h
@@ -17,7 +17,7 @@ struct cat {
 
     // Cat long-term properties
     enum cat_type type;
-    u32 time_per_frame;
+    u16 time_per_frame;
     u16 base_tile;
     u8 grid_x;
     u8 grid_y;
